+++ NEWS.md
@@ -1,6 +1,7 @@
 # fcl 0.0.7
 
 * Implement `bond_cf()` which give the forecasting cashflows of a bond.
+* Implement `create_rtn()` which allows for fast TWRR or Modified Return.
 
 # fcl 0.0.6